pc file: don't put -I/usr/include in cflags
If installed directly to a system header path, no need to add "I.." to
the compiler invocation (and we shouldn't).

Also, correctly handle shell-escaped variables (CXXOPTS, HEADERDIR,
etc). To do this export the variables and have the .pc file produced by
a shell script which can write the variables directly.

Doing this from inside the makefile is problematic.